browser/components/feeds/test/chrome/test_bug408328.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=408328
-->
<head>
  <title>Test feed preview safe-linkification</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=408328">Mozilla Bug 408328</a>
<p id="display"><iframe id="testFrame" src="http://mochi.test:8888/tests/browser/components/feeds/test/bug408328-data.xml"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 408328 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  var links = $("testFrame").contentDocument.getElementById("feedContent").getElementsByTagName("a");
  is(links.length, 5, "wrong number of linked items in feed preview");
  for (var i = 0; i < links.length; i++) {
    if (links[i].href)
      is(links[i].href, "http://example.org/first", "bad linkified item");
  }
});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
